Split widget: ui scope
